+++ uiautomator/src/main/java/android/support/test/uiautomator/UiDevice.java
@@ -2,6 +2,8 @@ package android.support.test.uiautomator;
 
 import android.app.Instrumentation;
 
+import java.io.File;
+
 public class UiDevice {
 
     public static UiDevice getInstance(Instrumentation context) {
@@ -10,5 +12,9 @@ public class UiDevice {
 
     public UiObject findObject(UiSelector selector) {
         return new UiObject();
+    }
+
+    public boolean takeScreenshot(File storePath) {
+        return false;
     }
 }